+++ lib/apathy_drive_web/channels/user_socket.ex
@@ -4,6 +4,8 @@ defmodule ApathyDriveWeb.UserSocket do
   ## Channels
   channel "mud:play", ApathyDriveWeb.MUDChannel
   channel "map", ApathyDriveWeb.MapChannel
+  channel "admin:abilities", ApathyDriveWeb.Admin.AbilitiesChannel
+  channel "admin:damage_types", ApathyDriveWeb.Admin.DamageTypesChannel
 
   ## Transports
   transport :websocket, Phoenix.Transports.WebSocket