+++ lib/apathy_drive_web/controllers/admin/damage_types_controller.ex
@@ -1,0 +1,8 @@
+defmodule ApathyDriveWeb.Admin.DamageTypesController do
+  use ApathyDrive.Web, :controller
+
+  def index(conn, _params) do
+    render conn, "index.html"
+  end
+
+end