+++ lib/apathy_drive_web/router.ex
@@ -1,6 +1,6 @@
 defmodule ApathyDriveWeb.Router do
   use Phoenix.Router
-  
+
   pipeline :browser do
     plug :accepts, ~w(html)
     plug :fetch_session
@@ -12,6 +12,7 @@ defmodule ApathyDriveWeb.Router do
 
   pipeline :admin do
     plug :require_admin
+    plug :put_layout, {ApathyDriveWeb.LayoutView, :admin}
   end
 
   scope "/", ApathyDriveWeb do
@@ -22,6 +23,14 @@ defmodule ApathyDriveWeb.Router do
     resources "/sessions", SessionController
     resources "/characters", CharacterController, only: [:create, :edit, :update],
                                                   singleton: true
+  end
+
+  scope "/admin", ApathyDriveWeb do
+    pipe_through :browser
+    pipe_through :admin
+
+    get "/abilities", Admin.AbilitiesController, :index
+    get "/damage-types", Admin.DamageTypesController, :index
   end
 
   # Fetch the current user from the session and add it to `conn.assigns`. This