+++ lib/apathy_drive_web/views/admin/damage_types_view.ex
@@ -1,0 +1,3 @@
+defmodule ApathyDriveWeb.Admin.DamageTypesView do
+  use ApathyDrive.Web, :view
+end