+++ priv/repo/migrations/20171028234808_add_damage_type_id_to_crits.exs
@@ -1,0 +1,9 @@
+defmodule ApathyDrive.Repo.Migrations.AddDamageTypeIdToCrits do
+  use Ecto.Migration
+
+  def change do
+    alter table(:crits) do
+      add :damage_type_id, references(:damage_types, on_delete: :delete_all)
+    end
+  end
+end