+++ priv/repo/migrations/20171029005533_add_messages_to_crits.exs
@@ -1,0 +1,11 @@
+defmodule ApathyDrive.Repo.Migrations.AddMessagesToCrits do
+  use Ecto.Migration
+
+  def change do
+    alter table(:crits) do
+      add :user_message, :text
+      add :target_message, :text
+      add :spectator_message, :text
+    end
+  end
+end